% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combineDatasets.R
\name{combineDatasets}
\alias{combineDatasets}
\title{Import data for pvGxE}
\usage{
combineDatasets(directory, lookFor = ".csv", verbose = T)
}
\arguments{
\item{directory}{The location of the csvs.}

\item{lookFor}{The character string to look for in the directory. Defaults
to ".csv"}

\item{verbose}{Should updates on progress be printed?}
}
\description{
\code{combineDatasets} Function to combine separate user-specified datasets
}

