% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getLineSiteMeans.R
\name{getLineSiteMeans}
\alias{getLineSiteMeans}
\title{Get genotype*site means.}
\usage{
getLineSiteMeans(dat, y)
}
\arguments{
\item{dat}{The dataframe containing parental data}

\item{y}{Character vector indicating the single response variable
Only 1 phenotype can be tested at a time.}
}
\description{
\code{getLineSiteMeans} Sub routine to get line means for the pvGxE Experiment
}

