% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaData.R
\docType{data}
\name{metaData}
\alias{metaData}
\title{The P. virgatum GxE experiment metadata}
\format{A data frame with 5617 rows and 11 variables}
\usage{
data(metaData)
\itemize{
  \item CROSS The cross origin of each line.
  \item LINE The 4-way or parental id.
  \item POP ...
  \item PLANT_ID Unique concatenation of CROSS, LINE, POP
  \item SITE Planting site
  \item PLOT_GL Global Plant ID - used for data merge
  \item PLOT_LC Local Plant ID
  \item CLM Column of planting design
  \item ROW Row of planting design
  \item NOTE_PLANTING ...
  \item NOTE_REFERENCE ...
}
}
\description{
A dataset containing the unique plant identifiers and all metadat.
}
\keyword{datasets}

