% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDistn.R
\name{plotDistn}
\alias{plotDistn}
\title{Get genotype*site means.}
\usage{
plotDistn(mergedData, yColumn, nbins = 30, histFill = "grey50",
  freeXAxis = F, freeYAxis = T, themeIn = theme_jtlbar, nrows = 3)
}
\arguments{
\item{yColumn}{Character vector indicating the single response variable}

\item{nbins}{The number of bins in the histogram}

\item{histFill}{The color for histogram fill}

\item{freeXAxis}{Logical, should the xaxis scales be independent}

\item{freeYAxis}{Logical, should the yaxis scales be independent}

\item{themeIn}{What ggplot theme should be used?}

\item{nrows}{The number of rows in the faceted panels.}

\item{dat}{The dataframe containing parental data}
}
\description{
\code{getLineSiteMeans} Sub routine to get line means for the pvGxE Experiment
}

