% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pvGxE.dataLoad.R
\name{pvGxE.dataLoad}
\alias{pvGxE.dataLoad}
\title{Import data for pvGxE}
\usage{
pvGxE.dataLoad(csv, directory = NULL, verbose = T)
}
\arguments{
\item{csv}{The name of the csv file.}

\item{directory}{The location of the csv.}
}
\description{
\code{pvGxE.dataLoad} Function to load user-supplied data into R.
}

